navigator/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
numpy>=1.24
pillow>=10.0
faiss-cpu>=1.8.0
torch>=2.2
torchvision>=0.17
timm>=0.9.16
tqdm>=4.66
pandas>=2.2
python-multipart>=0.0.9
scikit-image>=0.22
scipy>=1.10
